compilers/v1_0_6/requirements.txt
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
docutils==0.15.2
idna==2.8
pkginfo==1.5.0.1
Pygments==2.4.2
python-dotenv==0.11.0
PyYAML==3.12
readme-renderer==24.0
requests==2.22.0
requests-toolbelt==0.9.1
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.71
six==1.12.0
tqdm==4.33.0
twine==1.13.0
typing==3.7.4.1
urllib3==1.25.3
webencodings==0.5.1
